pyproject.toml
[project]
name = "mail-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "email-validator>=2.2.0",
  "fastapi[standard]>=0.115.12",
  "google-api-python-client>=2.166.0",
  "google-auth-oauthlib>=1.2.1",
  "jmespath>=1.0.1",
  "logfire>=4.14.2",
  "numpy>=2.2.4",
  "openpyxl>=3.1.5",
  "pendulum>=3.1.0",
  "platformdirs>=4.4.0",
  "psycopg2-binary>=2.9.10",
  "pydantic>=2.11.2",
  "pydantic-extra-types>=2.10.6",
  "pydantic-settings>=2.12.0",
  "pytest>=8.4.1",
  "sqlmodel>=0.0.24",
  "streamlit>=1.44.1",
]

[tool.uv.workspace]
members = ["fastapi"]

[dependency-groups]
dev = [
    "ipython>=9.3.0",
    "notebook>=7.4.5",
]
